Refs #18913: Apply rev suggestions
